Package structure, added unit tests
